requirements/development.txt
-r base.txt

pytest
pytest-django
pytest-cov
factory-boy
faker
ruff
black
isort
django-debug-toolbar
